% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_sce.R
\name{add_cell_qc}
\alias{add_cell_qc}
\title{Add scater's cell QC to the SCE}
\usage{
add_cell_qc(sce.object, is_mito)
}
\arguments{
\item{sce.object}{SCE object}

\item{is_mito}{vector of TRUE/FALSE indicating which rows correspond to mito-
chondrial genes}
}
\description{
Add scater's cell QC to the SCE
}
