% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_cdrFreqs.R
\name{add_chain_counts}
\alias{add_chain_counts}
\title{Add the occurrences of specific TRA/TRB chains to the colData of the SCE}
\usage{
add_chain_counts(sce.object)
}
\arguments{
\item{sce.object}{SCE object}
}
\description{
Add the occurrences of specific TRA/TRB chains to the colData of the SCE
}
