% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_cdrFreqs.R
\name{add_frequencies}
\alias{add_frequencies}
\title{Count the number of times individual cdr3s_aa entires occur}
\usage{
add_frequencies(
  sce.object,
  count_what = "cdr3s_aa",
  mode = c("all", "per.sample")
)
}
\arguments{
\item{sce.object}{SCE object}

\item{count_what}{Whether IDs or amino acid sequences should be counted.
Default: "cdr3s_aa"}

\item{mode}{either one of "all" or "per.sample" -- determines how
the number of individual cdr3s_aa entries is counted}
}
\value{
vector of numbers
}
\description{
Count the number of times individual cdr3s_aa entires occur
}
\examples{
\dontrun{
sce.filt$freq_per_Sample <- add_frequencies(sce.filt, mode = "per.sample")
sce.filt$freq_across_all <- add_frequencies(sce.filt, mode = "all")
}
}
