% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sharedClonotypes.R
\docType{data}
\name{clonotype_ids}
\alias{clonotype_ids}
\title{Table of customized clonotype IDs for sample MR4050}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 248 rows and 8 columns.
}
\usage{
clonotype_ids
}
\description{
\code{data.table} with our customized clonotype IDs for ease of
visualization and comparison. I.e., the CDR3s amino acid sequences are re-
placed with arbitrary IDs. Note thate these clonotypes are those that are
found in both conditions of patient MR4050, i.e. MUT and WT.
}
\details{
See section "Adding the clonotype IDs" in the vignette "Filtering and Processing"
about how the consolidation and clean up of the TRA/TRB sequences was done.
}
\seealso{
\code{shared_clonotypes}
}
\keyword{datasets}
