% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_qc.R
\name{do_gene_qc}
\alias{do_gene_qc}
\title{Identify genes to be kept}
\usage{
do_gene_qc(sce.object, min.cells = 5, return_df = FALSE)
}
\arguments{
\item{sce.object}{SingelCellExperiment object}

\item{min.cells}{minimum number of cells that must be expressing a given gene.
(Min. number of cells per Sample, that is). Default: 5}

\item{return_df}{whether the results of scater's QC should be returned in
a data.frame. Default: FALSE}
}
\value{
vector of TRUE/FALSE that indicates which genes passed the filters
and which ones did not
}
\description{
This function is a wrapper around scater::perFeatureQCMetrics
which is run on each sample (defined in the sce.object's "Sample" entry)
individually.
}
\examples{
\dontrun{
keep_genes <-  do_gene_qc(sce.filt)
sce.filt[keep_genes,]
}
}
