% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_markers}
\alias{extract_markers}
\title{Extract marker gene results from findMarkers into a data.table}
\usage{
extract_markers(
  sce.object,
  marker_search_result,
  FDR_thresh = 0.01,
  rank_thresh = 35,
  logFC_thresh = 0,
  comp_name = NULL
)
}
\arguments{
\item{sce.object}{SCE object}

\item{marker_search_result}{result of \code{scran::findMarkers}, e.g. l.mks$HFDvsWT;
typically a list where the names indicate the comparison. This function here
assumes that findMarkers was run with \code{direction = "up"}.}

\item{FDR_thresh}{Default: 0.01}

\item{rank_thresh}{maximum rank that the extracted genes must have in addition
to \code{FDR_thresh}}

\item{logFC_thresh}{minimum logFC that the extracted genes must have (default: 0)}

\item{comp_name}{add a name for the comparison that underlies the marker search
result, e.g. "DB_clst"; if NULL (default), just the name from the \code{marker_search_result}
will be used.}
}
\value{
data.table with three columns:
\itemize{
\item \code{gene_symbol}
\item \code{up_in}: in which condition(s) the gene was upregulated
\item \code{classify}: whether the gene was found in just one comparison or
multiple ones
}
}
\description{
Extract marker gene results from findMarkers into a data.table
}
\examples{
\dontrun{
gns_HFDvsWT <- extract_markers(sf, marker_search_result = l.mks$HFDvsWT,
                               FDR_thresh = 0.01, rank_thresh = 35)
}

}
\seealso{
\code{\link{plot_marker_heatmap}}
}
\author{
Friederike Dündar
}
