% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{fx.features_as_metaData}
\alias{fx.features_as_metaData}
\title{Features as meta data}
\usage{
fx.features_as_metaData(object, exprs_values, which_feat, meta.dt = NULL)
}
\arguments{
\item{object}{SCE object}

\item{exprs_values}{Types of expression values, i.e. which assay
to extract from (use \code{assayNames(sce)} if you're not sure
which ones to go with}

\item{which_feat}{feature name(s)}

\item{meta.dt}{data.table with already extracted metadata}
}
\description{
Extracts expression values for individual features and adds
them as a column
}
