% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\docType{data}
\name{gene_filt}
\alias{gene_filt}
\title{Filtered genes ' ' @format Named list of gene numbers following the filtering steps}
\format{
An object of class \code{list} of length 6.
}
\usage{
gene_filt
}
\description{
Filtered genes ' ' @format Named list of gene numbers following the filtering steps
}
\keyword{datasets}
