% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers_qc.R
\name{get_cc_info}
\alias{get_cc_info}
\title{Calculate cell cycle status per cell}
\usage{
get_cc_info(sce.object, print_results = TRUE)
}
\arguments{
\item{sce.object}{SCE object}

\item{print_results}{if set to TRUE (default), the total number of cells
per cell cycle phase is printed to the screen}
}
\value{
A list with \code{phases} (character vector of the predicted phase),
\code{scores} (data frame with the numeric phase scores for each phase and cell)
and \code{normalized.scores} (data frame with the row-normalized scores).
}
\description{
Wrapper function around \code{scran::cyclone()}. Assumes the
data is from a human data set.
}
\details{
Users should not filter out low-abundance genes before applying
cyclone. Even if a gene is not expressed in any cell, it may still be useful
for classification if it is phase-specific. Its lack of expression relative
to other genes will still yield informative pairs, and filtering them out
would reduce power.
}
\seealso{
\code{\link{scran::cyclone()}}
}
