% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_sce.R
\name{get_mito}
\alias{get_mito}
\title{Add chromosome information to the SCE object for each gene}
\usage{
get_mito(sce.object)
}
\description{
Add chromosome information to the SCE object for each gene
}
