% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{load_data_from_Box}
\alias{load_data_from_Box}
\title{Download and locally cache an RDA or RDS file that's stored on a website}
\usage{
load_data_from_Box(
 
    shared_link = "https://wcm.box.com/shared/static/mmf98o464n94fdh78x14sbwo7ts56ukq.rda",
  data_name = NULL,
  cache_path = NULL,
  check_for_update = FALSE,
  load_rds = TRUE
)
}
\arguments{
\item{shared_link}{The URL for the file as a string. NOTE: This should be the
\emph{direct} link, i.e. it should end with ".rda" or ".rdata", "rds" or the like}

\item{data_name}{A customized name that may be used to more easily refer to
the cached data (\code{rname} parameter of \code{BiocFileCache} functions),
e.g. "SCE2019" -- any kind of string will work, but it's important
to make a note of this name for reloading the data next time, otherwise the
data set will be downloaded again if there's a mismatch with the name.
If set to NULL (default), the \code{rname} in the cache will correspond to the
url indicated via \code{shared_link}.}

\item{cache_path}{If wanted, this allows to specify a path where the data
should be cached. if NULL (default), the default setting of \code{BiocFileCache()}
will be used.}

\item{check_for_update}{Indicate whether the data should be downloaded again.}
}
\value{
The final command here is \code{readRDS()}, so you should be assigning
the result of this function to an object name in your environment.
}
\description{
Using the library \code{BiocFileCache}, this function takes as
input the direct link to a RDA (or RDS) file, downloads it and stashes it in a
local directory from where it can be retrieved (using this exact function) in
a later session. I.e. if the data has already been downloaded, it won't be down-
loaded again.
}
\details{
Based on instructions from the BiocFileCache vignette: \url{https://bioconductor.org/packages/release/bioc/vignettes/BiocFileCache/inst/doc/BiocFileCache.html#local-cache-of-an-internet-resource}. This is a slightly more versatile form of
the function \code{load_RDSdata_from_Box()} as it also allows to load RDA files
directly into the global environment by setting the option \code{load_rds} to FALSE.
}
\examples{

## load RDS requires the user to define an object name (here: testingsce)
direct_link <- "https://wcm.box.com/shared/static/4887keeu77eskmdotve1sm7n6u7udckg.rds"
testingsce <- load_data_from_Box(shared_link = direct_link, data_name = "testRDS")

## loading RDA
sl <- "https://wcm.box.com/shared/static/mmf98o464n94fdh78x14sbwo7ts56ukq.rda"
load_data_from_Box(sl, load_rds = FALSE) # should add tiny_sce to global env.
ls()

}
