% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\name{load_MR4050filt}
\alias{load_MR4050filt}
\title{Load the SCE for the batch-corrected merged data set}
\usage{
sce.filt <- load_MR4050filt()
}
\value{
an SCE object that needs to be assigned to an object in the environment
}
\description{
Load the SingleCellExperiment object that holds the batch-corrected
reduced dimensionality results (MUT/WT = batch).
}
\seealso{
\code{\link{load_MR4050shared}}, \code{\link{load_MR4050merged}}
}
