% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\name{load_MR4050merged}
\alias{load_MR4050merged}
\title{Load the SCE for the batch-corrected merged data set}
\usage{
sce.merged <- load_MR4050merged()
}
\value{
an SCE object that needs to be assigned to an object in the environment
}
\description{
Load the SingleCellExperiment object that holds the batch-corrected
reduced dimensionality results (MUT/WT = batch).
}
