% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\name{load_MR4050shared}
\alias{load_MR4050shared}
\title{Load the SCE for the batch-corrected merged data set}
\usage{
sce.shared <- load_MR4050shared()
}
\value{
an SCE object that needs to be assigned to an object in the environment
}
\description{
Load the SingleCellExperiment object that holds the SCE
representing cells with clonotypes that are present in both conditions.
The UMAP coordinates were re-calculated on the reduced subset after removal
of suspected doublets (see the vignette about the filtering).
}
