% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sce.R
\name{load_sce}
\alias{load_sce}
\title{Load the filtered and processed SingleCellExperiment object}
\usage{
sce.filt <- load_sce(which_assays = "logcounts", sample = "MR4050")
}
\arguments{
\item{which_assays}{can be "all" (default) or individual assays, e.g. c("logcounts",
"counts") etc. If space and memory are problematic, definitely limit the selection here!
assays that are available are: "counts", "logcounts"}

\item{...}{Additional parameters passed on to \code{load_RDSdata_from_Box}, e.g.
\code{check_for_updates = TRUE}}
}
\value{
A SingleCellExperiment object with cells from 'mutant' samples
 (stimulation with tumor antigen) and from the 'wt' sample (stimulation
 with an irrelevant antigen)).

SCE object
}
\description{
Use this function to load the processed and filtered gene expression
data plus the clonotype information stored within one SingleCellExperiment
object.
}
\details{
For the entire code of the filtering and processing, see the vignette
\code{01_processing.Rmd}.

The resulting SCE object contains the usual content: colData with information
about individidual cells, rowData with info about individual genes, reducedDims,
etc.

The \code{colData} includes:

\describe{
\item{Barcode:}{Each cell's barcode used for keeping track of its identity during sequencing.}
\item{Sample:}{'WT' or 'MUT'}
\item{raw_clonotype_id:}{e.g. 'clonotype94'}
\item{cdr3s_aa:}{The amino acid sequence of the CDR3 portion, e.g. "TRA:CIARGGGGADGLTF;TRA:CGADRNGNEKLTF;TRB:CASSLTTDREPYEQYF"}
\item{multiTRA:}{TRUE/FALSE entries based on whether \code{cdr3s_aa} contained more than one entry for TRA}
\item{multiTRB:}{TRUE/FALSE entries based on whether \code{cdr3s_aa} contained more than one entry for TRB}
\item{numTRA:}{Number of TRA sequences within \code{cdr3s_aa}}
\item{numTRB:}{Number of TRB sequences within \code{cdr3s_aa}}
\item{cluster:}{clustering results of all cells}
}

The object also containes the coordinates from dimensionality reductions (see
examples for more details).
}
\examples{
\dontrun{

> library(SingleCellExperiment)
> sce.MR4050 <- load_sce(which_assays = "all", sample = "MR4050")

> reducedDimNames(sce.MR4050)
"corrected" "TSNE"      "UMAP"

> assayNames(sce.MR4050)
[1] "counts"                "logcounts"
}

}
