% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{make_long_dt}
\alias{make_long_dt}
\title{Prepare a data.table in skinny format suitable for ggplotting}
\usage{
make_long_dt(
  object,
  exprs_values = "logcounts",
  features = NULL,
  include_metaData = NULL
)
}
\arguments{
\item{object}{\code{SCE} object}

\item{exprs_values}{what type of expression values should be used from the
\code{SCE} object (default: "logcounts"); can be more than one name
(e.g., \code{names(assay(object))}).}

\item{features}{vector of feature names (e.g. gene names, HTOs, ADTs)
for which the \code{data.table} should be made
(optional; default: NULL)}

\item{include_metaData}{indicate whether the resulting \code{data.table}
should also include specified entries of colData(object). Cacn include specific
gene or other feature names. Default: NULL.}
}
\value{
\code{data.table} in skinny format with additional columns
"cell" and whatever was specified via \code{include_metaData}
}
\description{
Prepare a data.table in skinny format suitable for ggplotting
}
\examples{
\dontrun{
# get a long data.table with values from all assays
exp.dt <- make_long_dt(sce, exprs_values = names(assays(sce)),
                      features = c("SHOX2", "MYH6"),
                      include_metaData= c("SampleName", "Condition","Replicate"))
}

}
\seealso{
\code{\link{fx.extract_exprs}}
}
\author{
Friederike Dündar
}
