% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableprint.R
\name{my_table}
\alias{my_table}
\title{Function for printing pretty table}
\usage{
my_table(df, show_rownames = FALSE)
}
\arguments{
\item{show_rownames}{TRUE or FALSE (default)}

\item{data.frame}{data.frame to be printed}
}
\description{
Function for printing pretty table
}
