% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgrams.R
\name{plot_tgram}
\alias{plot_tgram}
\title{Plot THE PLOT}
\usage{
plot_tgram(
  in_dt = demks.dt,
  label_font_size = 12,
  gene_label_size = 16,
  dot_size = 1,
  dot_shape = 21,
  boxplot_stroke = 0.2,
  col_mut = "gray60",
  col_wt = "white",
  show_legend = TRUE
)
}
\arguments{
\item{in_dt}{result of \code{prep_data4tgram}}

\item{boxplot_stroke}{thickness of the boxplot outline, default: 0.2}
}
\value{
ggplot2 object
}
\description{
This function generates a plot that Chris Klebanoff calls
"Manhattan Plot". Basically, it will show the expression levels of the
gene(s) across different sample types and clonotypes. In addition, it
will add asterisks to those genes that were determined to be statistically
significantly different as provided in the results of \code{run_DE}.
}
\examples{
\dontrun{
delist <- lapply(ct$cdr3s_aa, function(x) run_DE(sce.filt2[, sce.filt2$cdr3s_aa == x],
                                                 group_identifier = "Sample",
                                                 direction = "up",
                                                 logfc = 0.2,
                                                 FDR = 0.05, rank = Inf,
                                                 comp_name = paste0(x, "_")))
names(delist) <- ct$cdr3s_aa
tt <- prep_data4tgram(sce.filt2, which_clonotypes = ct$cdr3s_aa,
                      DEresult_list = delist, goi = "IFNG")
plot_tgram(tt, dot_shape = 20, col_wt = "limegreen",col_mut = "dodgerblue1", label_font_size = 10)
}

}
\seealso{
\code{prep_data4tgram}, \code{run_DE}
}
