% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangling_sce.R
\name{prep_cell_info}
\alias{prep_cell_info}
\title{Add clonotype information to the colData of a SCE object}
\usage{
prep_cell_info(sce.object, clonotype_info = clono_info_wt)
}
\arguments{
\item{sce.object}{SCE object}

\item{clonotype_ino}{data.frame with the clonotype information for a given sample}
}
\value{
data.frame of the colData merged with the clonotype information
}
\description{
Add clonotype information to the colData of a SCE object
}
