% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgrams.R
\name{prep_data4tgram}
\alias{prep_data4tgram}
\title{Prepare data.table for plotting}
\usage{
prep_data4tgram(
  sce.object,
  which_clonotypes = NULL,
  DEresult_list,
  ignore_zeros = FALSE,
  goi,
  additional_colData = NULL
)
}
\arguments{
\item{sce.object}{SingleCellExperiment object where the colData should contain "Sample",
"cdr3s_aa", and "id" (= unique clonotype ID)}

\item{which_clonotypes}{vector of clonotype IDs that are going to be shown, e.g.
"C19"}

\item{DEresult_list}{list of findMarkers and extract_markers results for all
clonotypes present in \code{sce.object}}

\item{ignore_zeros}{set to TRUE to remove zeros from the expression data (Default: FALSE)}

\item{goi}{gene of interest for which the asterisks will be defined, e.g. "IFNG"}

\item{additional_colData}{additional information from the colData of the SCE to be added
to the returned data.table, e.g. "antigen".}
}
\value{
data.table
}
\description{
This function needs the clonotypes that are going to be plotted
and for which findMarkers results have been obtained via \code{run_DE}. It
will return a data.table with expression values (logcounts) for the gene of
interest (\code{goi}) as well as for CD4 and CD8A.
}
\examples{
\dontrun{
> genes_of_interest <- c("IFNG","CD4","CD8A","CD8B","CD3E","TNF","IL2","IL23A")
> tt <- prep_data4tgram(sce.shared,
    which_clonotypes = unique(sce.shared$id),
    DEresult_list = delist,
    goi = genes_of_interest)
> tt$Sample <- factor(tt$Sample, levels = c("wt","mut"), ordered = TRUE)
> plot_tgram(tt[gene_name == "IFNG"], dot_shape = 20, col_wt = "limegreen",
    col_mut = "dodgerblue1", label_font_size = 10, show_legend = FALSE)
}

}
\seealso{
\code{run_DE}, \code{plot_tgram}
}
