% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading_clonotypes.R
\name{reading_in_clonotypes}
\alias{reading_in_clonotypes}
\title{Reading in clonotype info per barcode}
\usage{
reading_in_clonotypes(cellRanger_result_dir)
}
\arguments{
\item{cellRanger_result_dir}{e.g. "/scratchLocal/frd2007/2018-11_Smita_Tcells/data/wildtype/".
 This should be the path to the un-tar'ed directory from cellRanger.

 @export}
}
\description{
Reading in clonotype info per barcode
}
\details{
This function expects the files "filtered_contig_annotations.csv" and
"clonotypes.csv" to be found in the "outs" folder after following \code{cellRanger_result_dir}.
These files are used to match the cell barcodes with unique clonotype IDs.
}
